doctests added for `range_func` and made `range_func.py`
